requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
python-jose==3.3.0
